requirements.txt
affine==2.4.0
albumentations==1.3.0
asttokens==2.4.1
attrs==24.2.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cjm-pytorch-utils==0.0.8
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cmake==3.31.1
comet-ml==3.47.3
comm==0.2.2
configobj==5.0.9
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
distlib==0.3.9
dulwich==0.22.6
einops==0.8.0
ensemble-boxes==1.0.9
everett==3.1.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
fire==0.7.0
fonttools==4.55.0
fsspec==2024.10.0
gmpy2==2.2.1
huggingface-hub==0.26.2
idna==3.10
imageio==2.37.0
intel-cmplr-lib-ur==2024.2.1
intel-openmp==2024.2.1
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
lit==18.1.8
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2024.2.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==9.5.0.50
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
opencv-python==4.6.0.66
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pybboxes==0.1.6
pycosat==0.6.6
pycparser==2.22
Pygments==2.18.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
PySocks==1.7.1
python-box==6.1.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qudida==0.0.4
rasterio==1.4.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
safetensors==0.4.5
sahi==0.11.21
scikit-image==0.25.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
sentry-sdk==2.19.0
shapely==2.0.6
simplejson==3.19.3
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
tbb==2021.13.1
termcolor==2.5.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2025.5.10
tokenizers==0.20.3
toolz==1.0.0
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchsummary==1.5.1
torchvision==0.16.1+cu118
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
triton==2.1.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.37
ultralytics-thop==2.0.12
urllib3==2.2.3
virtualenv==20.27.1
wcwidth==0.2.13
wrapt==1.17.0
wurlitzer==3.1.1
zstandard==0.23.0
